chore: Move notify to lib
