--- api.planx.uk/notify/index.ts
@@ -1,2 +1,0 @@
-export * from "./routeSendEmailRequest";
-export * from "./notify";
