+++ api.planx.uk/send/email.ts
@@ -2,7 +2,7 @@ import type { NextFunction, Request, Response } from "express";
 import { gql } from "graphql-request";
 import capitalize from "lodash/capitalize";
 import { markSessionAsSubmitted } from "../modules/saveAndReturn/service/utils";
-import { sendEmail } from "../notify";
+import { sendEmail } from "../lib/notify";
 import { EmailSubmissionNotifyConfig } from "../types";
 import { buildSubmissionExportZip } from "./exportZip";
 import { $api } from "../client";
